Agents_Course_Final_Assignment/requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
google-generativeai
python-dotenv
pandas
gradio
langchain-tavily